Switch "git add" to "git stage"
The two are aliases to the same operation. You could say that this
operation adds chunks/changes to the staging area, so both words "add"
and "stage" apply. Fair enough, but "stage" is a better choice. It makes
you aware of existence of the staging area; using git without being
aware of the staging area is a frustrating experience.